Refactor animations to utils and test blur bg for hero cards
